shortened messneger function in app.py
